requirements.txt
asn1crypto==0.24.0
astroid==2.0
boto3==1.7.52
botocore==1.10.52
CacheControl==0.12.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
Django==2.0.5
djangorestframework==3.8.2
docutils==0.14
Flask==1.0.2
idna==2.7
iso3166==0.8
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jwt==0.5.4
lazy-object-proxy==1.3.1
mapbox==0.16.1
mapboxcli==0.8.0
MarkupSafe==1.0
mccabe==0.6.1
msgpack==0.5.6
numpy==1.14.5
pandas==0.23.3
polyline==1.3.2
prompt-toolkit==2.0.4
pycparser==2.18
pylint==2.0.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
Werkzeug==0.14.1
wit==5.1.0
wrapt==1.10.11
